+++ NEWS.md
@@ -7,6 +7,7 @@ increments reflect backwards-incompatible API changes, not necessarily big chang
 * dplyr, knitr, rmarkdown and some other packages have moved to "Suggests:", lowering the dependency
   load considerably. dplyr verbs are still supported. huxtable gives an informative warning if 
   packages are not installed when needed.
+* New option "huxtable.knitr_output_format" to override the default output format in knitr documents.
 * Numeric row heights and column widths are rescaled to 1 when huxtables are cbinded/rbinded.
 * Bugfix: property accessors had the wrong environment. Thanks to Iñaki Úcar.
 * Bugfix: row heights and column widths weren't being copied with cbind/rbind.